Refactor site pages to utilize new content structure and improve metadata handling
- Replaced existing page components with new content components for Home, Authors, Monthly Review, News, Privacy Policy, Terms of Service, and Tags.
- Centralized metadata generation for improved SEO and consistency across pages.
- Streamlined imports and component structure for better maintainability and readability.
- Enhanced the overall organization of the site by implementing a modular approach to page content.
