+++ src/app/(site)/terms/page.tsx
@@ -1,177 +1,5 @@
-import { LegalDocPage } from '@/components/blocks';
-import { LEGAL_LAST_UPDATED } from '@/config/legal';
-import { buildMetadata } from '@/lib/seo';
+import { termsPageMetadata, TermsPageContent } from '@/components/modules/legal';
 
-export const metadata = buildMetadata({
-  title: 'Terms of Service',
-  description:
-    'Terms governing your use of ACTA News, our editorial and publication services, and related features.',
-  path: '/terms',
-});
+export const metadata = termsPageMetadata;
 
-export default function TermsOfServicePage() {
-  return (
-    <LegalDocPage title="Terms of Service" lastUpdated={LEGAL_LAST_UPDATED}>
-      <p>
-        These Terms of Service (&ldquo;Terms&rdquo;) form a legal agreement between you and ACTA
-        (&ldquo;ACTA,&rdquo; &ldquo;we,&rdquo; &ldquo;us,&rdquo; or &ldquo;our&rdquo;) regarding
-        your access to and use of the ACTA News website, associated applications, and related
-        services (collectively, the &ldquo;Services&rdquo;). If you do not agree, do not use the
-        Services.
-      </p>
-      <p>
-        These Terms describe important rights and obligations. We encourage you to read them
-        carefully and, where appropriate, to obtain independent legal advice. Nothing here creates a
-        partnership, agency, or joint venture.
-      </p>
-
-      <h2>1. Acceptance and changes</h2>
-      <p>
-        By creating an account, accessing, or using the Services, you agree to these Terms and our
-        Privacy Policy. We may change these Terms from time to time. We will post the updated Terms
-        on this page and update the &ldquo;Last updated&rdquo; date. Your continued use after
-        changes become effective constitutes acceptance of the revised Terms. If you do not agree,
-        you must stop using the Services.
-      </p>
-
-      <h2>2. The Services</h2>
-      <p>
-        ACTA News publishes editorial and informational content, including news, articles, and
-        periodic reviews. We may add, modify, or discontinue features, content, or integrations. The
-        Services are provided on an &ldquo;as is&rdquo; and &ldquo;as available&rdquo; basis to the
-        maximum extent permitted by law.
-      </p>
-
-      <h2>3. Eligibility</h2>
-      <p>
-        You may use the Services only if you can form a binding contract with ACTA under applicable
-        law. If you use the Services on behalf of an organization, you represent that you have
-        authority to bind that organization, and &ldquo;you&rdquo; includes the organization. You
-        must not use the Services if you are prohibited from doing so by applicable sanctions,
-        export, or other laws.
-      </p>
-
-      <h2>4. Accounts and access</h2>
-      <p>
-        Certain features may require an account. You are responsible for maintaining the
-        confidentiality of your credentials, for all activity under your account, and for providing
-        accurate information. You must notify us promptly of any unauthorized use. We may suspend or
-        terminate access if we reasonably believe you have violated these Terms, pose a security
-        risk, or if required by law. Authentication flows (including social sign-in) may be provided
-        by third parties; their use is also subject to their terms and policies.
-      </p>
-
-      <h2>5. Acceptable use</h2>
-      <p>You agree not to:</p>
-      <ul>
-        <li>Violate any law, regulation, or third-party rights</li>
-        <li>
-          Attempt to gain unauthorized access to the Services, systems, or other users&rsquo; data
-        </li>
-        <li>
-          Interfere with or disrupt the Services, including by transmitting malware, spam, or
-          harmful code
-        </li>
-        <li>
-          Scrape, index, or collect data in bulk without our prior written consent, where such use
-          is prohibited
-        </li>
-        <li>Impersonate any person or entity, or misrepresent your affiliation</li>
-        <li>
-          Use the Services to build a competing product or to reverse engineer the Services, except
-          where such restriction is unenforceable under applicable law
-        </li>
-      </ul>
-
-      <h2>6. Content and intellectual property</h2>
-      <h3>Our content</h3>
-      <p>
-        Text, graphics, logos, and other materials made available by ACTA through the Services
-        (excluding third-party or user content as identified) are owned by ACTA or our licensors and
-        are protected by copyright, trademark, and other laws. You may not copy, modify, distribute,
-        or create derivative works except as allowed by us in writing or as permitted for sharing
-        via built-in features (e.g. links and ordinary browser viewing).
-      </p>
-      <h3>Your content</h3>
-      <p>
-        If you submit content to us (e.g. comments, contributions, or materials for publication),
-        you grant ACTA a worldwide, non-exclusive, royalty-free license to use, host, reproduce,
-        modify, display, and distribute such content for the purpose of operating, promoting, and
-        improving the Services. You represent that you have the rights to grant this license. You
-        retain ownership of your content subject to the license above.
-      </p>
-
-      <h2>7. Third-party links and services</h2>
-      <p>
-        The Services may link to or integrate third-party sites, services, or content. We do not
-        control and are not responsible for third parties. Your use of third-party services is at
-        your own risk and subject to their terms and privacy practices.
-      </p>
-
-      <h2>8. Disclaimers</h2>
-      <p>
-        THE SERVICES AND ALL CONTENT ARE PROVIDED &ldquo;AS IS&rdquo; AND &ldquo;AS AVAILABLE&rdquo;
-        WITHOUT WARRANTIES OF ANY KIND, WHETHER EXPRESS, IMPLIED, OR STATUTORY, INCLUDING WARRANTIES
-        OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE, TITLE, AND NON-INFRINGEMENT. ACTA DOES
-        NOT WARRANT THAT THE SERVICES WILL BE UNINTERRUPTED, ERROR FREE, OR FREE OF HARMFUL
-        COMPONENTS. EDITORIAL CONTENT IS FOR INFORMATIONAL PURPOSES; IT IS NOT LEGAL, FINANCIAL,
-        INVESTMENT, OR PROFESSIONAL ADVICE.
-      </p>
-
-      <h2>9. Limitation of liability</h2>
-      <p>
-        TO THE FULLEST EXTENT PERMITTED BY APPLICABLE LAW, IN NO EVENT WILL ACTA, ITS AFFILIATES,
-        OFFICERS, DIRECTORS, EMPLOYEES, OR AGENTS BE LIABLE FOR ANY INDIRECT, INCIDENTAL, SPECIAL,
-        CONSEQUENTIAL, OR PUNITIVE DAMAGES, OR ANY LOSS OF PROFITS, DATA, GOODWILL, OR BUSINESS
-        OPPORTUNITY, ARISING OUT OF OR RELATED TO YOUR USE OF THE SERVICES, WHETHER IN CONTRACT,
-        TORT, OR OTHERWISE, EVEN IF WE HAVE BEEN ADVISED OF THE POSSIBILITY OF SUCH DAMAGES.
-      </p>
-      <p>
-        OUR TOTAL LIABILITY FOR ANY CLAIMS ARISING OUT OF OR RELATING TO THE SERVICES OR THESE TERMS
-        WILL NOT EXCEED THE GREATER OF (A) THE AMOUNT YOU PAID US FOR THE SERVICES DURING THE TWELVE
-        (12) MONTHS BEFORE THE CLAIM, OR (B) FIFTY U.S. DOLLARS (US $50.00), IF THE SERVICES WERE
-        FREE TO USE. SOME JURISDICTIONS DO NOT ALLOW CERTAIN LIMITATIONS; IN THOSE JURISDICTIONS,
-        OUR LIABILITY IS LIMITED TO THE MAXIMUM EXTENT PERMITTED BY LAW.
-      </p>
-
-      <h2>10. Indemnity</h2>
-      <p>
-        You will defend, indemnify, and hold harmless ACTA and its affiliates, officers, directors,
-        and employees from and against any claims, damages, losses, or expenses (including
-        reasonable attorneys&rsquo; fees) arising out of your use of the Services, your content, or
-        your violation of these Terms or applicable law, except to the extent caused by our gross
-        negligence or willful misconduct.
-      </p>
-
-      <h2>11. Governing law; dispute resolution</h2>
-      <p>
-        These Terms are governed by the laws of the State of California and the United States,
-        without regard to conflict of law rules. The United Nations Convention on Contracts for the
-        International Sale of Goods does not apply. Subject to applicable law, the exclusive
-        jurisdiction and venue for disputes arising from these Terms or the Services will be the
-        state and federal courts located in San Francisco County, California, and you consent to
-        personal jurisdiction in those courts. If you are a consumer, mandatory consumer protections
-        in your country of residence may still apply to you; nothing in this section limits those
-        rights.
-      </p>
-
-      <h2>12. General</h2>
-      <p>
-        These Terms, together with our Privacy Policy, constitute the entire agreement between you
-        and ACTA regarding the Services. If a provision is held invalid, the remainder remains in
-        effect. Our failure to enforce a right is not a waiver. You may not assign these Terms
-        without our consent; we may assign them in connection with a merger, acquisition, or sale of
-        assets. Section headings are for convenience only.
-      </p>
-
-      <h2>13. Contact</h2>
-      <p>
-        For questions about these Terms, contact us through the channels provided on our website at{' '}
-        <a href="https://acta.build" rel="noreferrer" target="_blank">
-          acta.build
-        </a>
-        .
-      </p>
-    </LegalDocPage>
-  );
-}
+export default TermsPageContent;
