+++ src/app/login/page.tsx
@@ -1,16 +1,5 @@
-import { AuthPage } from '@/components/blocks';
-import { buildMetadata } from '@/lib/seo';
-import { siteConfig } from '@/config/site';
+import { LoginPageContent, loginPageMetadata } from '@/components/modules/auth';
 
-export const metadata = buildMetadata({
-  title: 'Sign in',
-  description: `Sign in to ${siteConfig.name} (preview UI).`,
-  path: '/login',
-});
+export const metadata = loginPageMetadata;
 
-/**
- * UI-only sign-in (Efferd auth-5). Renders without the public site header/footer; see `app/(site)/layout.tsx`.
- */
-export default function LoginPage() {
-  return <AuthPage />;
-}
+export default LoginPageContent;
