+++ src/components/modules/tags/index.ts
@@ -1,3 +1,4 @@
 export { TagBadge } from './ui/TagBadge';
 export { TagCloud } from './ui/TagCloud';
 export { fetchTags, fetchTagBySlug } from './services/tags.service';
+export { TagPageContent, generateTagMetadata } from './pages/TagPage';
